EIP1-1920 Refactor to allow manual control of container lifecycle
